Add files via upload
The Greatest Among Three program is designed to find the largest number among three given integers. It uses conditional statements (if-else if-else) to compare the values of the three integers. Depending on the comparisons, the program determines which number is the greatest and stores it in the greatest variable. Finally, it displays the greatest number on the console. This program is a practical example of decision-making in Java and demonstrates how to use logical operators (&&) for multiple conditions. It can be extended for more numbers or modified for other similar comparisons.