Add files via upload
# Leap Year Checker

This Java program determines whether a given year is a leap year. By taking a year input from the user, it checks the conditions for leap years: divisible by 4 but not by 100, or divisible by 400. If the year meets these criteria, the program outputs "It is a leap year"; otherwise, it outputs "It is not a leap year". This project serves as a simple yet effective demonstration of basic programming concepts, including conditionals and user input. It's perfect for beginners looking to understand fundamental Java skills. Feel free to explore and build upon this program!