Add files via upload
### AddRandomNumbers.java

This Java program generates two random numbers between 0 and 99, adds them together, and prints the result. The Random class is utilized to generate the random numbers, and the sum is calculated and displayed in the console. A simple and effective demonstration of basic random number generation and arithmetic operations in Java.