Add files via upload
### AddTwoNumbers.java

This Java program prompts the user to input two numbers, calculates their sum, and displays the result. It utilizes the Scanner class to read user input from the console. A straightforward example of basic user interaction and arithmetic operations in Java, perfect for beginners to understand how to capture and process input data.