Add files via upload
##Factorial Calculator in Java

This Java program computes the factorial of a user-provided number. It employs the Scanner class to read the user input. The factorial calculation is performed using a for loop that multiplies integers from 1 up to the given number. Once the computation is complete, the result is displayed to the user. This program is straightforward and efficient, making it a useful tool for learning basic Java programming and iterative processes.