Add files via upload
### Multiple of 27 Checker in Java

This program in Java helps you determine whether a given number is a multiple of 27. By employing the Scanner class, it takes user input and checks the number using a modulus operation. Depending on the result, it outputs an appropriate message to the user. It's a simple and efficient way to demonstrate basic input handling and conditional logic in Java.