Add files via upload
##LeapYearChecker 
This Java program checks whether a given year is a *leap year* or not. It takes user input, applies the leap year conditions, and determines if February has *28 or 29 days. 
The logic follows standard leap year rules based on divisibility by **4, 100, and 400*. Simple and efficient, this program ensures accurate results for any input year. 🚀